+++ pkg/util/validation/limits_test.go
@@ -3,6 +3,7 @@ package validation
 import (
 	"encoding/json"
 	"reflect"
+	"strings"
 	"testing"
 	"time"
 
@@ -165,6 +166,15 @@ func TestLimitsLoadingFromJson(t *testing.T) {
 
 	assert.Equal(t, 0.5, l.IngestionRate, "from json")
 	assert.Equal(t, 100, l.MaxLabelNameLength, "from defaults")
+
+	// Unmarshal should fail if input contains unknown struct fields and
+	// the decoder flag `json.Decoder.DisallowUnknownFields()` is set
+	inp = `{"unknown_fields": 100}`
+	l = Limits{}
+	dec := json.NewDecoder(strings.NewReader(inp))
+	dec.DisallowUnknownFields()
+	err = dec.Decode(&l)
+	assert.Error(t, err)
 }
 
 func TestLimitsTagsYamlMatchJson(t *testing.T) {